+++ src/components/ConnectedTasksJobs/constants.js
@@ -1,0 +1,4 @@
+export const CONNECTED_RECORD_TYPES = {
+  AGREEMENT: 'agreement',
+  AGREEMENT_LINE: 'agreementLine',
+};